+++ src/templates/objects_profile/oyProfile_s_.template.c
@@ -49,7 +49,7 @@
              );
     else
       sprintf( &text[strlen(text)], "%lu",
-               s->size_
+               (long unsigned int)s->size_
              );
   } else
   if((int)type >= oyNAME_DESCRIPTION)
